ml-engine/chatbot/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
sentence-transformers
chromadb
python-dotenv
